feat!: dockerize collab and refactor room and editor files in frontend
