+++ docker-compose.yml
@@ -16,6 +16,7 @@ services:
       NEXT_PUBLIC_USER_SERVICE_PORT: 8001
       NEXT_PUBLIC_MATCHING_SERVICE_PORT: 8002
       NEXT_PUBLIC_QUESTION_SERVICE_PORT: 8003
+      NEXT_PUBLIC_COLLABORATION_SERVICE_PORT: 8004
     networks:
       - cs3219-project-network
 
@@ -50,6 +51,19 @@ services:
     networks:
       - cs3219-project-network
 
+  collaboration-service:
+    container_name: collaboration-service
+    build:
+      context: ./backend
+      dockerfile: Dockerfile.collaboration
+    volumes:
+      - ./backend/collaboration-service:/app/backend/matching-service
+      - /app/backend/collaboration-service/node_modules
+    ports:
+      - '8004:8004'
+    networks:
+      - cs3219-project-network
+
   question-service:
     container_name: question-service
     depends_on:
@@ -61,7 +75,7 @@ services:
       - ./backend/question-service:/app/backend/question-service
       - /app/backend/question-service/node_modules
     ports:
-      - "8003:8003"
+      - '8003:8003'
     environment:
       DB_LOCAL_URI: mongodb://mongo_db:27017/question
     networks:
