--- frontend/components/CodeEditor/index.ts
@@ -1,2 +1,0 @@
-import CodeEditor from './CodeEditor';
-export default CodeEditor;
